donate project

+++ src/main/java/fptu/fcharity/service/WalletService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Service
@@ -15,7 +16,7 @@ public class WalletService {
     @Transactional
     public Wallet save() {
         Wallet wallet = new Wallet();
-        wallet.setBalance(0);
+        wallet.setBalance(BigDecimal.valueOf(0.00));
         return walletRepository.save(wallet);
     }
 }
