+++ src/main/java/fptu/fcharity/service/manage/organization/OrganizationServiceImpl.java
@@ -17,6 +17,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
@@ -79,7 +80,7 @@ public class OrganizationServiceImpl implements OrganizationService {
         organization.setCeo(ceo);
 
         Wallet wallet = new Wallet();
-        wallet.setBalance(0);
+        wallet.setBalance(BigDecimal.valueOf(0));
         Wallet savedWallet =  walletRepository.save(wallet);
         organization.setWalletAddress(savedWallet);
         Organization organizationSaved = organizationRepository.save(organization);
